feat: Expose workflow and dm methods in snake_case. (#2492)
* feat: Expose workflow and dm methods in snake_case.

* Add additional methods as well.

* Update docs.

* Minor fixes.

* Update src/ansys/fluent/core/workflow.py


* Update src/ansys/fluent/core/workflow.py


* workflow fixes.

* Update tests.

* Minor fixes.

* Update test.

* Added rename and tests for the same.

---------
