+++ src/ansys/fluent/core/workflow.py
@@ -12,6 +12,7 @@ import ansys.fluent.core as pyfluent
 from ansys.fluent.core.data_model_cache import DataModelCache
 from ansys.fluent.core.services.datamodel_se import (
     PyCallableStateObject,
+    PyCommand,
     PyMenuGeneric,
     PySingletonCommandArgumentsSubItem,
 )
@@ -320,10 +321,6 @@ class BaseTask:
                 pass
         return self._python_name
 
-    def delete(self) -> None:
-        """Delete this task from the workflow."""
-        self._command_source.DeleteTasks(ListOfTasks=[self.name()])
-
     def _get_camel_case_arg_keys(self):
         _args = self.arguments
         _camel_args = []
@@ -373,20 +370,61 @@ class BaseTask:
             for arg in [*self.arguments().keys(), *dir(self._task)]:
                 arg_list.append(arg)
 
-        return list(
-            dict.fromkeys(
-                sorted(set(list(self.__dict__.keys()) + dir(type(self)) + arg_list))
-            )
-        )
+        return sorted(set(list(self.__dict__.keys()) + dir(type(self)) + arg_list))
+
+    def delete(self) -> None:
+        """Delete this task from the workflow."""
+        self._command_source.delete_tasks(list_of_tasks=[self.python_name()])
+
+    def rename(self, new_name: str):
+        """Rename the current task with the name passed as argument."""
+        return self._task.Rename(NewName=new_name)
 
     def add_child_to_task(self):
+        """Add a child task."""
         return self._task.AddChildToTask()
 
     def update_child_tasks(self, setup_type_changed: bool):
+        """Update child tasks."""
         self._task.UpdateChildTasks(SetupTypeChanged=setup_type_changed)
 
     def insert_compound_child_task(self):
+        """Insert a compound child task."""
         return self._task.InsertCompoundChildTask()
+
+    def get_next_possible_tasks(self) -> list[str]:
+        """Get the list of possible names of commands that can be inserted as tasks
+        after this current task is executed."""
+        return [camel_to_snake_case(task) for task in self._task.GetNextPossibleTasks()]
+
+    def insert_next_task(self, command_name: str):
+        """Insert a task based on the command name passed as argument after the current
+        task is executed.
+
+        Parameters
+        ----------
+        command_name: str
+            Name of the new task.
+
+        Returns
+        -------
+        None
+
+        Raises
+        ------
+        ValueError
+            If command_name does not match a task name.
+        """
+        if command_name not in self.get_next_possible_tasks():
+            raise ValueError(
+                f"'{command_name}' cannot be inserted next to '{self.python_name()}'. \n"
+                "Please use 'get_next_possible_tasks()' to view list of allowed tasks."
+            )
+        return self._task.InsertNextTask(
+            CommandName=snake_to_camel_case(
+                command_name, self._task.GetNextPossibleTasks()
+            )
+        )
 
     def __call__(self, **kwds) -> Any:
         if kwds:
@@ -840,6 +878,11 @@ class CompositeTask(BaseTask):
         """
         return {}
 
+    def insert_composite_child_task(self, command_name: str):
+        """Insert a composite child task based on the command name passed as
+        argument."""
+        return self._task.InsertCompositeChildTask(CommandName=command_name)
+
 
 class ConditionalTask(CommandTask):
     """Conditional task representation for wrapping a Workflow TaskObject instance of
@@ -901,8 +944,12 @@ class CompoundTask(CommandTask):
             Optional state.
         """
         state = state or {}
-        state.update({"AddChild": "yes"})
-        self._task.Arguments.set_state(state)
+        if self._dynamic_interface:
+            state.update({"add_child": "yes"})
+            self.arguments.set_state(state)
+        else:
+            state.update({"AddChild": "yes"})
+            self._task.Arguments.set_state(state)
 
     def add_child_and_update(self, state=None):
         """Add a child to this CompoundTask and update.
@@ -1002,6 +1049,22 @@ class NewWorkflowWrapper:
         self._main_thread_ident = None
         self._task_objects = {}
         self._dynamic_interface = False
+        self._help_string_command_id_map = {}
+        self._help_string_display_text_map = {}
+        self._unwanted_attrs = {
+            "reset_workflow",
+            "initialize_workflow",
+            "load_workflow",
+            "create_new_workflow",
+            # "fault_tolerant",
+            # "part_management",
+            # "pm_file_management",
+            "rules",
+            "service",
+            "task_object",
+            # "watertight",
+            "workflow",
+        }
 
     def task(self, name: str) -> BaseTask:
         """Get a TaskObject by name, in a BaseTask wrapper. The wrapper adds extra
@@ -1092,7 +1155,7 @@ class NewWorkflowWrapper:
         _task_object = self._task_objects.get(attr)
         if _task_object:
             return _task_object
-        if attr != "TaskObject":
+        if attr != "TaskObject" and attr not in self._unwanted_attrs:
             if not attr.islower():
                 raise AttributeError(
                     "Camel case attribute access is not supported. "
@@ -1109,18 +1172,14 @@ class NewWorkflowWrapper:
         """Override the behaviour of dir to include attributes in WorkflowWrapper and
         the underlying workflow."""
         arg_list = [camel_to_snake_case(arg) for arg in dir(self._workflow)]
-        return list(
-            dict.fromkeys(
-                sorted(
-                    set(
-                        list(self.__dict__.keys())
-                        + dir(type(self))
-                        + arg_list
-                        + self.child_task_python_names()
-                    )
-                )
-            )
+        dir_set = set(
+            list(self.__dict__)
+            + dir(type(self))
+            + arg_list
+            + self.child_task_python_names()
         )
+        dir_set = dir_set - self._unwanted_attrs
+        return sorted(filter(None, dir_set))
 
     def __call__(self):
         """Delegate calls to the underlying workflow."""
@@ -1177,6 +1236,128 @@ class NewWorkflowWrapper:
                 self._refreshing = False
 
             self.add_on_affected(refresh_after_sleep)
+
+    def save_workflow(self, file_path: str):
+        """Save the current workflow to the location provided."""
+        self._workflow.SaveWorkflow(FilePath=file_path)
+
+    def load_state(self, list_of_roots: list):
+        """Load the state of the workflow."""
+        self._workflow.LoadState(ListOfRoots=list_of_roots)
+
+    def _populate_help_string_command_id_map(self):
+        if not self._help_string_command_id_map:
+            for command in dir(self._command_source):
+                if command in ["SwitchToSolution", "set_state"]:
+                    continue
+                command_obj = getattr(self._command_source, command)
+                if isinstance(command_obj, PyCommand):
+                    command_obj_instance = command_obj.create_instance()
+                    help_str = command_obj_instance.get_attr("helpString")
+                    if help_str and help_str.islower():
+                        self._help_string_command_id_map[help_str] = command
+                        self._help_string_display_text_map[
+                            help_str
+                        ] = command_obj_instance.get_attr("displayText")
+                    del command_obj_instance
+
+    def get_possible_tasks(self):
+        """Get the list of possible names of commands that can be inserted as tasks."""
+        self._populate_help_string_command_id_map()
+        return list(self._help_string_command_id_map)
+
+    def insert_new_task(self, command_name: str):
+        """Insert a new task based on the command name passed as argument.
+
+        Parameters
+        ----------
+        command_name: str
+            Name of the new task.
+
+        Returns
+        -------
+        None
+
+        Raises
+        ------
+        ValueError
+            If command_name does not match a task name.
+        """
+        self._populate_help_string_command_id_map()
+        if command_name not in self._help_string_command_id_map:
+            raise ValueError(
+                f"'{command_name}' is not an allowed command task.\n"
+                "Please use 'get_possible_tasks()' to view list of allowed command tasks."
+            )
+        return self._workflow.InsertNewTask(
+            CommandName=self._help_string_command_id_map[command_name]
+        )
+
+    def delete_tasks(self, list_of_tasks: list[str]):
+        """Delete the list of tasks passed as argument.
+
+        Parameters
+        ----------
+        list_of_tasks: list[str]
+            List of task items.
+
+        Returns
+        -------
+        None
+
+        Raises
+        ------
+        ValueError
+            If command_name does not match a task name. None of the tasks is deleted.
+        """
+        self._populate_help_string_command_id_map()
+        list_of_tasks_with_display_name = []
+        for task_name in list_of_tasks:
+            try:
+                list_of_tasks_with_display_name.append(
+                    self._help_string_display_text_map[task_name]
+                )
+            except KeyError as ex:
+                raise ValueError(
+                    f"'{task_name}' is not an allowed command task.\n"
+                    "Please use 'get_possible_tasks()' to view list of allowed command tasks."
+                ) from ex
+
+        return self._workflow.DeleteTasks(ListOfTasks=list_of_tasks_with_display_name)
+
+    def create_composite_task(self, list_of_tasks: list[str]):
+        """Create the list of tasks passed as argument.
+
+        Parameters
+        ----------
+        list_of_tasks: list[str]
+            List of task items.
+
+        Returns
+        -------
+        None
+
+        Raises
+        ------
+        RuntimeError
+            If command_name does not match a task name.
+        """
+        self._populate_help_string_command_id_map()
+        list_of_tasks_with_display_name = []
+        for task_name in list_of_tasks:
+            try:
+                list_of_tasks_with_display_name.append(
+                    self._help_string_display_text_map[task_name]
+                )
+            except KeyError:
+                raise RuntimeError(
+                    f"'{task_name}' is not an allowed command task.\n"
+                    "Please use 'get_possible_tasks()' to view list of allowed command tasks."
+                )
+
+        return self._workflow.CreateCompositeTask(
+            ListOfTasks=list_of_tasks_with_display_name
+        )
 
 
 class OldWorkflowWrapper:
