+++ tests/test_new_meshing_workflow.py
@@ -1,8 +1,10 @@
+import time
 from typing import Iterable
 
 import pytest
 
 from ansys.fluent.core import examples
+from ansys.fluent.core.meshing.watertight import watertight_workflow
 
 
 @pytest.mark.nightly
@@ -527,25 +529,394 @@ def test_workflow_and_data_model_methods_new_meshing_workflow(new_mesh_session):
         "mixing_elbow.pmdb", "pyfluent/mixing_elbow"
     )
     watertight = meshing.watertight()
-    assert (
-        meshing.workflow.CreateCompositeTask.__class__
-        == watertight.create_composite_task.__class__
+
+    # Checks if any of the unwanted attrs are present in dir call
+    assert (set(dir(watertight)) - watertight._unwanted_attrs) == set(dir(watertight))
+
+    for attr in watertight._unwanted_attrs:
+        with pytest.raises(AttributeError) as msg:
+            getattr(watertight, attr)
+        assert (
+            msg.value.args[0]
+            == f"'NewMeshingWorkflow' object has no attribute '{attr}'"
+        )
+
+    watertight.import_geometry.rename(new_name="import_geom_wtm")
+    assert len(watertight._task_list) == 11
+    watertight.insert_new_task("import_geometry")
+    assert len(watertight._task_list) == 12
+    watertight.task("import_geom_wtm").file_name = import_file_name
+    watertight.task("import_geom_wtm").length_unit = "in"
+    watertight.task("import_geom_wtm")()
+    assert watertight.task("import_geom_wtm").get_next_possible_tasks() == [
+        "import_body_of_influence_geometry",
+        "set_up_periodic_boundaries",
+        "create_local_refinement_regions",
+        "run_custom_journal",
+    ]
+    watertight.task("import_geom_wtm").insert_next_task(
+        "import_body_of_influence_geometry"
     )
-    assert (
-        meshing.workflow.InsertNewTask.__class__ == watertight.insert_new_task.__class__
+    watertight.task("import_geom_wtm").insert_next_task("set_up_periodic_boundaries")
+    assert len(watertight._task_list) == 14
+
+
+@pytest.mark.fluent_version(">=23.2")
+@pytest.mark.codegen_required
+def test_watertight_workflow(mixing_elbow_geometry, new_mesh_session):
+    watertight = watertight_workflow(
+        geometry_file_name=mixing_elbow_geometry, session=new_mesh_session
     )
-    assert (
-        meshing.workflow.TaskObject["Import Geometry"].__class__
-        == watertight.import_geometry.__class__
+    add_local_sizing = watertight.add_local_sizing
+    assert not add_local_sizing.ordered_children()
+    add_local_sizing._add_child(state={"boi_face_label_list": ["cold-inlet"]})
+    assert not add_local_sizing.ordered_children()
+    added_sizing = add_local_sizing.add_child_and_update(
+        state={"boi_face_label_list": ["elbow-fluid"]}
     )
-    assert (
-        meshing.workflow.TaskObject["Import Geometry"].GetNextPossibleTasks()
-        == watertight.import_geometry.get_next_possible_tasks()
+    assert len(add_local_sizing.ordered_children()) == 1
+    assert added_sizing
+    assert added_sizing.boi_face_label_list() == ["elbow-fluid"]
+
+
+@pytest.mark.fluent_version(">=23.2")
+@pytest.mark.codegen_required
+def test_watertight_workflow_children(mixing_elbow_geometry, new_mesh_session):
+    watertight = watertight_workflow(
+        geometry_file_name=mixing_elbow_geometry, session=new_mesh_session
     )
-    assert (
-        meshing.workflow.TaskObject["Import Geometry"].InsertNextTask.__class__
-        == watertight.import_geometry.insert_next_task.__class__
+    add_local_sizing = watertight.add_local_sizing
+    assert not add_local_sizing.ordered_children()
+    add_local_sizing._add_child(state={"boi_face_label_list": ["cold-inlet"]})
+    assert not add_local_sizing.ordered_children()
+    added_sizing = add_local_sizing.add_child_and_update(
+        state={"boi_face_label_list": ["elbow-fluid"]}
     )
-    watertight.import_geometry.file_name = import_file_name
-    watertight.import_geometry.length_unit = "in"
-    watertight.import_geometry()
+    assert len(add_local_sizing.ordered_children()) == 1
+    assert added_sizing
+    assert added_sizing.boi_face_label_list() == ["elbow-fluid"]
+    assert added_sizing.name() == "facesize_1"
+    assert len(added_sizing.arguments())
+    added_sizing_by_name = add_local_sizing.compound_child("facesize_1")
+    added_sizing_by_pos = add_local_sizing.last_child()
+    assert added_sizing.arguments() == added_sizing_by_name.arguments()
+    assert added_sizing.arguments() == added_sizing_by_pos.arguments()
+    assert not added_sizing.python_name()
+    describe_geometry = watertight.describe_geometry
+    describe_geometry_children = describe_geometry.ordered_children()
+    assert len(describe_geometry_children) == 2
+    describe_geometry_child_task_python_names = (
+        describe_geometry.child_task_python_names()
+    )
+    assert describe_geometry_child_task_python_names == [
+        "enclose_fluid_regions",
+        "create_regions",
+    ]
+
+
+@pytest.mark.fluent_version(">=23.2")
+@pytest.mark.codegen_required
+def test_watertight_workflow_dynamic_interface(mixing_elbow_geometry, new_mesh_session):
+    watertight = watertight_workflow(
+        geometry_file_name=mixing_elbow_geometry, session=new_mesh_session
+    )
+    create_volume_mesh = watertight.create_volume_mesh
+    assert create_volume_mesh is not None
+    watertight.delete_tasks(list_of_tasks=["create_volume_mesh"])
+    # I assume that what's going on here is that due to DeleteTasks we are triggering
+    # change events in the server but those events are (still) being transmitted after
+    # DeleteTasks has returned. Hence, the dynamic watertight Python interface
+    # is still updating after the command has returned and the client can try to access
+    # while it is in that update phase, leading to (difficult to understand) exceptions.
+    # Temporarily sleeping in the test. I note that the core event tests use sleeps also.
+    time.sleep(2.5)
+    with pytest.raises(AttributeError) as msg:
+        watertight.create_volume_mesh
+    assert (
+        msg.value.args[0]
+        == "'NewMeshingWorkflow' object has no attribute 'create_volume_mesh'"
+    )
+    watertight.insert_new_task(command_name="create_volume_mesh")
+    time.sleep(2.5)
+    create_volume_mesh = watertight.create_volume_mesh
+    assert create_volume_mesh is not None
+
+    watertight_geom = watertight.describe_geometry
+    assert watertight_geom.create_regions.arguments()["number_of_flow_volumes"] == 1
+    watertight.delete_tasks(list_of_tasks=["create_regions"])
+    assert watertight_geom.create_regions is None
+    assert watertight_geom.enclose_fluid_regions
+    watertight_geom.enclose_fluid_regions.delete()
+    assert watertight_geom.enclose_fluid_regions is None
+    watertight.create_volume_mesh.delete()
+    time.sleep(2.5)
+    with pytest.raises(AttributeError) as msg:
+        watertight.create_volume_mesh
+    assert (
+        msg.value.args[0]
+        == "'NewMeshingWorkflow' object has no attribute 'create_volume_mesh'"
+    )
+
+
+@pytest.mark.fluent_version(">=23.2")
+@pytest.mark.codegen_required
+def test_extended_wrapper(new_mesh_session, mixing_elbow_geometry):
+    watertight = new_mesh_session.watertight()
+    import_geometry = watertight.import_geometry
+    assert import_geometry.Arguments() == {}
+    import_geometry.Arguments = dict(FileName=mixing_elbow_geometry)
+    assert 8 < len(import_geometry.arguments.get_state()) < 15
+    assert len(import_geometry.arguments.get_state(explicit_only=True)) == 1
+    import_geometry.arguments.set_state(dict(file_name=None))
+    time.sleep(5)
+    assert import_geometry.arguments.get_state(explicit_only=True) == dict(
+        file_name=None
+    )
+    assert import_geometry.arguments.get_state()["file_name"] is None
+    import_geometry.arguments.set_state(dict(file_name=mixing_elbow_geometry))
+    time.sleep(5)
+    assert import_geometry.arguments.get_state(explicit_only=True) == dict(
+        file_name=mixing_elbow_geometry
+    )
+    assert import_geometry.file_name() == mixing_elbow_geometry
+    import_geometry.file_name.set_state("bob")
+    time.sleep(5)
+    assert import_geometry.file_name() == "bob"
+    import_geometry.file_name.set_state(mixing_elbow_geometry)
+    import_geometry()
+    add_local_sizing = watertight.add_local_sizing
+    assert not add_local_sizing.ordered_children()
+    # new_mesh_session.workflow.TaskObject["Add Local Sizing"]._add_child(state={"BOIFaceLabelList": ["elbow-fluid"]})
+    add_local_sizing._add_child(state={"boi_face_label_list": ["cold-inlet"]})
+    assert not add_local_sizing.ordered_children()
+
+    added_sizing = add_local_sizing.add_child_and_update(
+        state={"boi_face_label_list": ["elbow-fluid"]}
+    )
+    assert len(add_local_sizing.ordered_children()) == 1
+    assert added_sizing
+    assert added_sizing.boi_face_label_list() == ["elbow-fluid"]
+    # restart
+    watertight = new_mesh_session.watertight()
+    assert import_geometry.state() == "Out-of-date"
+    import_geometry(FileName=mixing_elbow_geometry, AppendMesh=False)
+    assert import_geometry.state() == "Up-to-date"
+    import_geometry_state = import_geometry.arguments()
+    assert len(import_geometry_state) > 2
+
+
+@pytest.mark.fluent_version(">=23.1")
+@pytest.mark.skip
+def test_meshing_workflow_structure(new_mesh_session):
+    """
+    o Workflow
+    |
+    |--o Import Geometry
+    |
+    |--o Add Local Sizing
+    |
+    |--o Generate the Surface Mesh
+    |
+    |--o Describe Geometry
+    |  |
+    |  |--o Enclose Fluid Regions (Capping)
+    |  |
+    |  |--o Create Regions
+    |
+    |--o Update Regions
+    |
+    |--o Add Boundary Layers
+    |
+    |--o Generate the Volume Mesh
+    """
+    w = new_mesh_session.workflow
+    w.InitializeWorkflow(WorkflowType="Watertight Geometry")
+
+    task_names = (
+        "Import Geometry",
+        "Add Local Sizing",
+        "Generate the Surface Mesh",
+        "Describe Geometry",
+        "Enclose Fluid Regions (Capping)",
+        "Create Regions",
+        "Update Regions",
+        "Add Boundary Layers",
+        "Generate the Volume Mesh",
+    )
+
+    (
+        import_geom,
+        add_sizing,
+        gen_surf_mesh,
+        describe_geometry,
+        cap,
+        create_regions,
+        update_regions,
+        add_boundary_layers,
+        gen_vol_mesh,
+    ) = all_tasks = [w.task(name) for name in task_names]
+
+    def upstream_names(task):
+        return {upstream.name() for upstream in task.get_direct_upstream_tasks()}
+
+    def downstream_names(task):
+        return {downstream.name() for downstream in task.get_direct_downstream_tasks()}
+
+    assert upstream_names(import_geom) == set()
+    assert downstream_names(import_geom) == {
+        "Generate the Surface Mesh",
+        "Add Local Sizing",
+    }
+
+    assert upstream_names(add_sizing) == {"Import Geometry"}
+    assert downstream_names(add_sizing) == {"Generate the Surface Mesh"}
+
+    assert upstream_names(gen_surf_mesh) == {"Import Geometry", "Add Local Sizing"}
+    assert downstream_names(gen_surf_mesh) == {
+        "Describe Geometry",
+        "Add Boundary Layers",
+        "Generate the Volume Mesh",
+    }
+
+    assert upstream_names(describe_geometry) == {
+        "Generate the Surface Mesh",
+        "Add Boundary Layers",
+    }
+    assert downstream_names(describe_geometry) == {
+        "Update Regions",
+        "Add Boundary Layers",
+        "Generate the Volume Mesh",
+    }
+
+    assert upstream_names(cap) == {
+        "Describe Geometry",
+        "Add Boundary Layers",
+        "Generate the Surface Mesh",
+    }
+    assert downstream_names(cap) == {
+        "Describe Geometry",
+        "Add Boundary Layers",
+        "Generate the Volume Mesh",
+    }
+
+    assert upstream_names(create_regions) == {
+        "Describe Geometry",
+        "Add Boundary Layers",
+        "Generate the Surface Mesh",
+    }
+    assert downstream_names(create_regions) == {
+        "Describe Geometry",
+        "Add Boundary Layers",
+        "Generate the Volume Mesh",
+        "Update Regions",
+    }
+
+    assert upstream_names(update_regions) == {"Describe Geometry"}
+    assert downstream_names(update_regions) == {"Generate the Volume Mesh"}
+
+    assert upstream_names(add_boundary_layers) == {
+        "Describe Geometry",
+        "Generate the Surface Mesh",
+    }
+    assert downstream_names(add_boundary_layers) == {
+        "Describe Geometry",
+        "Generate the Volume Mesh",
+    }
+
+    assert upstream_names(gen_vol_mesh) == {
+        "Update Regions",
+        "Describe Geometry",
+        "Add Boundary Layers",
+        "Generate the Surface Mesh",
+    }
+    assert downstream_names(gen_vol_mesh) == set()
+
+    for task in all_tasks:
+        assert {sub_task.name() for sub_task in task.ordered_children()} == (
+            {
+                "Enclose Fluid Regions (Capping)",
+                "Create Regions",
+            }
+            if task is describe_geometry
+            else set()
+        )
+
+    for task in all_tasks:
+        assert {sub_task.name() for sub_task in task.inactive_ordered_children()} == (
+            {
+                "Apply Share Topology",
+                "Update Boundaries",
+            }
+            if task is describe_geometry
+            else set()
+        )
+
+    task_ids = [task.get_id() for task in all_tasks]
+    # uniqueness test
+    assert len(set(task_ids)) == len(task_ids)
+    # ordering test
+    idxs = [int(id[len("TaskObject") :]) for id in task_ids]
+    assert sorted(idxs) == idxs
+    """Given the workflow::
+
+            Workflow
+            ├── Import Geometry
+            ├── Add Local Sizing
+            ├── Generate the Surface Mesh ── Insert Next Task
+                                            ├── Add Boundary Type
+                                            ├── Update Boundaries
+                                            ├── ...
+    """
+    assert set(gen_surf_mesh.GetNextPossibleTasks()) == {
+        "AddBoundaryType",
+        "UpdateBoundaries",
+        "SetUpPeriodicBoundaries",
+        "LinearMeshPattern",
+        "ManageZones",
+        "ModifyMeshRefinement",
+        "ImproveSurfaceMesh",
+        "RunCustomJournal",
+    }
+
+    children = w.ordered_children()
+    expected_task_order = (
+        "Import Geometry",
+        "Add Local Sizing",
+        "Generate the Surface Mesh",
+        "Describe Geometry",
+        "Update Regions",
+        "Add Boundary Layers",
+        "Generate the Volume Mesh",
+    )
+
+    actual_task_order = tuple(child.name() for child in children)
+
+    assert actual_task_order == expected_task_order
+
+    assert [child.name() for child in children[3].ordered_children()] == [
+        "Enclose Fluid Regions (Capping)",
+        "Create Regions",
+    ]
+
+    gen_surf_mesh.InsertNextTask(CommandName="AddBoundaryType")
+
+    children = w.ordered_children()
+    expected_task_order = (
+        "Import Geometry",
+        "Add Local Sizing",
+        "Generate the Surface Mesh",
+        "Add Boundary Type",
+        "Describe Geometry",
+        "Update Regions",
+        "Add Boundary Layers",
+        "Generate the Volume Mesh",
+    )
+
+    actual_task_order = tuple(child.name() for child in children)
+
+    assert actual_task_order == expected_task_order
+
+    assert [child.name() for child in children[4].ordered_children()] == [
+        "Enclose Fluid Regions (Capping)",
+        "Create Regions",
+    ]
